+++ compiler/src/main/scala/org/scalaexercises/exercises/compiler/comments.scala
@@ -3,8 +3,7 @@
  * Copyright (C) 2015-2016 47 Degrees, LLC. <http://www.47deg.com>
  */
 
-package org.scalaexercises.exercises
-package compiler
+package org.scalaexercises.compiler
 
 import scala.language.higherKinds
 
@@ -21,7 +20,7 @@ import scala.xml.Xhtml
 import cats.{ Id, Eq, Functor }
 import cats.data.Xor
 
-import org.scalaexercises.exercises.formatting._
+import org.scalaexercises.compiler.formatting._
 
 /** Facade for the different layers of comment processing. */
 object Comments {
