+++ definitions/src/main/scala/exercise/Section.scala
@@ -3,7 +3,7 @@
  * Copyright (C) 2015-2016 47 Degrees, LLC. <http://www.47deg.com>
  */
 
-package exercise
+package org.scalaexercises.definitions
 
 /** Marker trait for exercise sections.
   */
