+++ runtime/src/main/scala/org/scalaexercises/exercises/Security.scala
@@ -1,3 +1,8 @@
+/*
+ * scala-exercises-runtime
+ * Copyright (C) 2015-2016 47 Degrees, LLC. <http://www.47deg.com>
+ */
+
 package org.scalaexercises.runtime
 
 import java.security._
