+++ runtime/src/test/scala/com/fortysevendeg/exercises/LibraryDiscoverySpec.scala
@@ -3,7 +3,9 @@
  * Copyright (C) 2015-2016 47 Degrees, LLC. <http://www.47deg.com>
  */
 
-package org.scalaexercises.exercises
+package org.scalaexercises.runtime
+
+import org.scalaexercises.runtime.model.Exercise
 
 import org.scalatest._
